ts version tested for get, post driver/rider
