+++ nodeAppPostPg/index.ts
@@ -9,10 +9,11 @@ const GoodFile    = require('good-file');
 const config      = require('./dbInfo.js');
 const logOptions  = require('./logInfo.js');
 
-const dbDefs      = require('./dbDefs.js');
-const dbQueries   = require('./dbQueries.js')
+const dbQueries   = require('./dbQueries.js');
 const logging     = require('./logging.js');
 const postgresQueries = require('./postgresQueries.js');
+
+const routeFns = require('./routeFunctions.js');
 
 config.user       = process.env.PGUSER;
 config.database   = process.env.PGDATABASE;
@@ -24,6 +25,8 @@ config.port       = process.env.PGPORT;
 // not passing config causes Client() to search for env vars
 const pool = new Pool();
 const server = new Hapi.Server();
+
+routeFns.setPool(pool);
 
 const OPS_INTERVAL  = 300000; // 5 mins
 const DEFAULT_PORT  = process.env.PORT || 3000;
@@ -101,18 +104,6 @@ server.route({
   }
 });
 
-function getResultStrings(tableName) {
-    var resultStrings = {
-      success: ' row inserted',
-      failure: ' row insert failed' 
-    }
-
-    resultStrings.success = tableName + resultStrings.success; 
-    resultStrings.failure = tableName + resultStrings.failure; 
-
-    return resultStrings;
-}
-
 function getExecResultStrings(tableName) {
     var resultStrings = {
       success: ' fn called: ',
@@ -125,85 +116,27 @@ function getExecResultStrings(tableName) {
     return resultStrings;
 }
 
-function sanitiseDriver(payload) {
-  if (payload.DriverCollectionRadius === undefined ||
-      payload.DriverCollectionRadius === "") {
-    // console.log("santising...");
-    payload.DriverCollectionRadius = 0;
-  }    
-}
-
 server.route({
   method: 'POST',
-  path: '/' + dbDefs.DRIVER_ROUTE,
-  handler: (req, reply) => {
-    var payload = req.payload;
-    var results = getResultStrings(dbDefs.DRIVER_ROUTE);
-
-    console.log("driver radius1 : " + payload.DriverCollectionRadius);
-    sanitiseDriver(payload);
-    console.log("driver radius2 : " + payload.DriverCollectionRadius);
-
-    req.log();
-
-    console.log("driver payload: " + JSON.stringify(payload, null, 4));
-    console.log("driver zip: " + payload.DriverCollectionZIP);
-
-    postgresQueries.dbInsertData(payload, pool, dbQueries.dbGetInsertDriverString, 
-                  getDriverPayloadAsArray,
-                  req, reply, results);
-  }
-});
-
-function sanitiseRider(payload) {
-
-  if (payload.RiderVotingState === undefined) {
-    payload.RiderVotingState = "MO";
-  }
-}
-
-server.route({
-  method: 'POST',
-  path: '/' + dbDefs.RIDER_ROUTE,
-  handler: (req, reply) => {
-    var payload = req.payload;
-    var results = getResultStrings(dbDefs.RIDER_ROUTE);
-
-    console.log("rider state1 : " + payload.RiderVotingState);
-    sanitiseRider(payload);
-    console.log("rider state2 : " + payload.RiderVotingState);
-
-    req.log();
-
-    console.log("rider payload: " + JSON.stringify(payload, null, 4));
-    console.log("rider zip: " + payload.RiderCollectionZIP);
-
-    postgresQueries.dbInsertData(payload, pool, dbQueries.dbGetInsertRiderString, 
-                  getRiderPayloadAsArray,
-                  req, reply, results);
-  }
+  path: '/' + routeFns.DRIVER_ROUTE,
+  handler: routeFns.postDriver
 });
 
 server.route({
   method: 'POST',
-  path: '/' + dbDefs.HELPER_ROUTE,
-  handler: (req, reply) => {
-    var payload = req.payload;
-    var results = getResultStrings(dbDefs.HELPER_ROUTE);
+  path: '/' + routeFns.RIDER_ROUTE,
+  handler: routeFns.postRider
+});
 
-    req.log();
-
-    console.log("helper payload: " + JSON.stringify(payload, null, 4));
-
-    postgresQueries.dbInsertData(payload, pool, dbQueries.dbGetInsertHelperString, 
-                  getHelperPayloadAsArray,
-                  req, reply, results);
-  }
+server.route({
+  method: 'POST',
+  path: '/' + routeFns.HELPER_ROUTE,
+  handler: routeFns.postHelper
 });
 
 server.route({
   method: 'DELETE',
-  path: '/' + dbDefs.DELETE_ROUTE,
+  path: '/' + routeFns.DELETE_ROUTE,
   handler: (req, reply) => {
     var payload = req.payload;
     var results = getExecResultStrings('cancel ride: ');
@@ -220,7 +153,7 @@ server.route({
 
 server.route({
   method: 'PUT',
-  path: '/' + dbDefs.PUT_ROUTE,
+  path: '/' + routeFns.PUT_ROUTE,
   handler: (req, reply) => {
     var payload = req.payload;
     var results = getExecResultStrings('reject ride: ');
@@ -263,13 +196,6 @@ server.register({
 
 logging.logReqResp(server, pool);
 
-function getHelperPayloadAsArray(req, payload) {
-  return [      
-        payload.Name, payload.Email, payload.Capability,
-        1, moment().toISOString()
-    ]
-}
-
 function getRejectRidePayloadAsArray(req, payload) {
   return [
         payload.UUID
@@ -282,36 +208,3 @@ function getCancelRidePayloadAsArray(req, payload) {
     ]
 }
 
-function getRiderPayloadAsArray(req, payload) {
-  return [      
-        req.info.remoteAddress, payload.RiderFirstName, payload.RiderLastName, payload.RiderEmail
-      , payload.RiderPhone, payload.RiderVotingState
-      , payload.RiderCollectionZIP, payload.RiderDropOffZIP, payload.AvailableRideTimesJSON
-      , payload.TotalPartySize
-      , (payload.TwoWayTripNeeded ? 'true' : 'false')
-      , payload.RiderPreferredContactMethod
-      , (payload.RiderIsVulnerable ? 'true' : 'false')
-      , (payload.RiderWillNotTalkPolitics ? 'true' : 'false')
-      , (payload.PleaseStayInTouch ? 'true' : 'false')
-      , (payload.NeedWheelchair ? 'true' : 'false')
-      , payload.RiderAccommodationNotes
-      , (payload.RiderLegalConsent ? 'true' : 'false')
-    ]
-}
-
-function getDriverPayloadAsArray(req, payload) {
-  return [
-        req.info.remoteAddress, payload.DriverCollectionZIP, payload.DriverCollectionRadius, payload.AvailableDriveTimesJSON
-      , (payload.DriverCanLoadRiderWithWheelchair ? 'true'  : 'false')
-      , payload.SeatCount
-      , (payload.DriverHasInsurance ? 'true' : 'false')
-      , payload.DriverFirstName, payload.DriverLastName
-      , payload.DriverEmail, payload.DriverPhone
-      , (payload.DrivingOnBehalfOfOrganization ? 'true' : 'false')
-      , payload.DrivingOBOOrganizationName 
-      , (payload.RidersCanSeeDriverDetails ? 'true' : 'false')
-      , (payload.DriverWillNotTalkPolitics ? 'true' : 'false')
-      , (payload.PleaseStayInTouch ? 'true' : 'false')
-      , payload.DriverLicenceNumber 
-    ]
-}
